Refactor component imports and delete unused files
